+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 
+## [1.0.2] - 2025-09-18
+### Fixed
+- Updated committee images.
+- Testing contact email form.
+
 ## [1.0.1] - 2024-09-23
 ### Fixed
 - Updated committee images.
 - Link changed to EventBrite instead of DOCX files.
 - Links to Bootstrap changed to CDN.
-
 
 ## [1.0.0] - 2024-08-07
 ### Added